now renders videos. might be issue with pixel aspect ratio though
TODO: create relevant entries in database and copy renders back to server.
TODO: check for files that aren't in_use with no session_id and remove
them
TODO: Check if aspect ration of output files is correct. The widths are
coming out less then the (correct) heights which is strange.