Started work on creating hls and dash encodes when the output mp4's are large
This is because with long files there is quite a lot of metadata and can
cause the video to take a long time to load in browsers as all the
metadata has to be retrieved first before the user can seek anywhere.
Providing hls and dash encodes (ios, web browsers), fixes this issue as
these won't be the large amounts of meatadata for these. The mp4's will
still be available as fallbacks.